@@ -1,6 +1,6 @@
 //
 //  Request.swift
-//  Swiftcord
+//  Sword
 //
 //  Created by Alejandro Alonso
 //  Copyright © 2017 Alejandro Alonso. All rights reserved.
@@ -15,236 +15,132 @@ import Dispatch
 /// HTTP Handler
 extension Swiftcord {
 
-  // MARK: Functions
+    /**
+     Actual HTTP Request
 
-  /**
-   Actual HTTP Request
+     - parameter url: URL to request
+     - parameter params: Optional URL Query Parameters to send
+     - parameter body: Optional Data to send to server
+     - parameter file: Optional for when files
+     - parameter authorization: Whether or not the Authorization header is required by Discord
+     - parameter method: Type of HTTP Method
+     - parameter rateLimited: Whether or not the HTTP request needs to be rate limited
+     - parameter reason: Optional for when user wants to specify audit-log reason
+     */
+    func request(
+        _ endpoint: Endpoint,
+        params: [String: Any]? = nil,
+        body: [String: Any]? = nil,
+        file: String? = nil,
+        authorization: Bool = true,
+        rateLimited: Bool = true,
+        reason: String? = nil
+    ) async throws -> Any? {
+        let endpointInfo = endpoint.httpInfo
 
-   - parameter url: URL to request
-   - parameter params: Optional URL Query Parameters to send
-   - parameter body: Optional Data to send to server
-   - parameter file: Optional for when files
-   - parameter authorization: Whether or not the Authorization header is required by Discord
-   - parameter method: Type of HTTP Method
-   - parameter rateLimited: Whether or not the HTTP request needs to be rate limited
-   - parameter reason: Optional for when user wants to specify audit-log reason
-  */
-  func request(
-    _ endpoint: Endpoint,
-    params: [String: Any]? = nil,
-    body: [String: Any]? = nil,
-    file: String? = nil,
-    authorization: Bool = true,
-    rateLimited: Bool = true,
-    reason: String? = nil,
-    then completion: @escaping (Any?, RequestError?) -> Void
-  ) {
-    let sema = DispatchSemaphore(value: 0)
+        var route = self.getRoute(for: endpointInfo.url)
 
-    let endpointInfo = endpoint.httpInfo
+        if route.hasSuffix("/messages/:id") && endpointInfo.method == .delete {
+            route += ".delete"
+        }
 
-    var route = self.getRoute(for: endpointInfo.url)
+        var urlString = "https://discord.com/api/v9\(endpointInfo.url)"
 
-    if route.hasSuffix("/messages/:id") && endpointInfo.method == .delete {
-      route += ".delete"
-    }
+        if let params = params {
+            urlString += "?"
+            urlString += params.map({ key, value in "\(key)=\(value)" }
+            ).joined(separator: "&")
+        }
 
-    var urlString = "https://discord.com/api/v9\(endpointInfo.url)"
+        guard let url = URL(string: urlString) else {
+            self.error(
+                "Used an invalid URL: \"\(urlString)\". Please report this."
+            )
+            return nil
+        }
 
-    if let params = params {
-      urlString += "?"
-      urlString += params.map({ key, value in "\(key)=\(value)" }
-      ).joined(separator: "&")
-    }
+        var request = URLRequest(url: url)
 
-    guard let url = URL(string: urlString) else {
-      self.error(
-        "[Swiftcord] Used an invalid URL: \"\(urlString)\". Please report this."
-      )
-      return
-    }
+        request.httpMethod = endpointInfo.method.rawValue
 
-    var request = URLRequest(url: url)
+        if authorization {
+            if self.options.isBot {
+                request.addValue("Bot \(token)", forHTTPHeaderField: "Authorization")
+            } else {
+                request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+            }
+        }
 
-    request.httpMethod = endpointInfo.method.rawValue
+        if let reason = reason {
+            request.addValue(
+                reason.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!,
+                forHTTPHeaderField: "X-Audit-Log-Reason"
+            )
+        }
 
-    if authorization {
-      if self.options.isBot {
-        request.addValue("Bot \(token)", forHTTPHeaderField: "Authorization")
-      } else {
-        request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-      }
-    }
-
-    if let reason = reason {
-      request.addValue(
-        reason.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!,
-        forHTTPHeaderField: "X-Audit-Log-Reason"
-      )
-    }
-
-    request.addValue(
-      "DiscordBot (https://github.com/SketchMaster2001/Swiftcord, 0.9.0)",
-      forHTTPHeaderField: "User-Agent"
-    )
-
-    if let body = body {
-      if let array = body["array"] as? [Any] {
-        request.httpBody = array.createBody()
-      } else {
-        request.httpBody = body.createBody()
-      }
-
-      request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-    }
-
-    #if os(macOS)
-    if let file = file {
-      let boundary = createBoundary()
-
-      let payloadJson: String?
-
-      if let array = body?["array"] as? [Any] {
-        payloadJson = array.encode()
-      } else {
-        payloadJson = body?.encode()
-      }
-
-      request.httpBody = try? self.createMultipartBody(
-        with: payloadJson,
-        fileUrl: file, boundary: boundary
-      )
-      request.addValue(
-        "multipart/form-data; boundary=\(boundary)",
-        forHTTPHeaderField: "Content-Type"
-      )
-    }
-    #endif
-
-    let task = self.session.dataTask(with: request) {
-      [unowned self, unowned sema] data, response, error in
-
-      let response = response as! HTTPURLResponse
-      let headers = response.allHeaderFields
-
-      if error != nil {
-        #if !os(Linux)
-        completion(nil, RequestError(error! as NSError))
-        #else
-        completion(nil, RequestError(error as! NSError))
-        #endif
-        sema.signal()
-        return
-      }
-
-      if rateLimited {
-        self.handleRateLimitHeaders(
-          headers["x-ratelimit-limit"],
-          headers["x-ratelimit-reset"],
-          (headers["Date"] as! String).httpDate.timeIntervalSince1970,
-          route
+        request.addValue(
+            "DiscordBot (https://github.com/SketchMaster2001/Swiftcord, 1.0.0)",
+            forHTTPHeaderField: "User-Agent"
         )
-      }
 
-      if response.statusCode == 204 {
-        completion(nil, nil)
-        sema.signal()
-        return
-      }
-
-      let returnedData = try? JSONSerialization.jsonObject(
-        with: data!,
-        options: .allowFragments
-      )
-
-      if response.statusCode != 200 && response.statusCode != 201 {
-
-        if response.statusCode == 429 {
-          print(
-            "[Swiftcord] You're being rate limited. (This shouldn't happen, check your system clock)"
-          )
-
-          let retryAfter = Int(headers["retry-after"] as! String)!
-          let global = headers["x-ratelimit-global"] as? Bool
-
-          guard global == nil else {
-            self.isGloballyLocked = true
-            self.globalQueue.asyncAfter(
-              deadline: DispatchTime.now() + .seconds(retryAfter)
-            ) { [unowned self] in
-              self.globalUnlock()
+        if let body = body {
+            if let array = body["array"] as? [Any] {
+                request.httpBody = array.createBody()
+            } else {
+                request.httpBody = body.createBody()
             }
 
-            sema.signal()
-            return
-          }
-
-          self.globalQueue.asyncAfter(
-            deadline: DispatchTime.now() + .seconds(retryAfter)
-          ) { [unowned self] in
-            self.request(
-              endpoint,
-              body: body,
-              file: file,
-              authorization: authorization,
-              rateLimited: rateLimited,
-              then: completion
-            )
-          }
+            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
         }
 
-        if response.statusCode >= 500 {
-          self.globalQueue.asyncAfter(
-            deadline: DispatchTime.now() + .seconds(3)
-          ) { [unowned self] in
-            self.request(
-              endpoint,
-              body: body,
-              file: file,
-              authorization: authorization,
-              rateLimited: rateLimited,
-              then: completion
+        #if os(macOS)
+        if let file = file {
+            let boundary = createBoundary()
+
+            let payloadJson: String?
+
+            if let array = body?["array"] as? [Any] {
+                payloadJson = array.encode()
+            } else {
+                payloadJson = body?.encode()
+            }
+
+            request.httpBody = try? self.createMultipartBody(
+                with: payloadJson,
+                fileUrl: file, boundary: boundary
             )
-          }
-
-          sema.signal()
-          return
+            request.addValue(
+                "multipart/form-data; boundary=\(boundary)",
+                forHTTPHeaderField: "Content-Type"
+            )
         }
+        #endif
 
-        completion(nil, RequestError(response.statusCode, returnedData!))
-        sema.signal()
-        return
-      }
+        var returnData: Any?
+        returnData = try await withCheckedThrowingContinuation({ continuation in
+            self.baseRequest(
+                request,
+                endpoint,
+                route: route,
+                params: params,
+                body: body,
+                file: file,
+                authorization: authorization,
+                rateLimited: rateLimited,
+                reason: reason
+            ) { data, err in
+                if let err = err {
+                    continuation.resume(throwing: err)
+                }
 
-      completion(returnedData, nil)
+                if data != nil {
+                    continuation.resume(returning: data)
+                }
+            }
+        })
 
-      sema.signal()
+        return returnData
     }
-
-    let apiCall = { [unowned self] in
-      guard rateLimited, self.rateLimits[route] != nil else {
-        task.resume()
-
-        sema.wait()
-        return
-      }
-
-      let item = DispatchWorkItem {
-        task.resume()
-
-        sema.wait()
-      }
-
-      self.rateLimits[route]!.queue(item)
-    }
-
-    if !self.isGloballyLocked {
-      apiCall()
-    } else {
-      self.globalRequestQueue.append(apiCall)
-    }
-
-  }
 
     /**
      Actual HTTP Request
@@ -257,195 +153,376 @@ extension Swiftcord {
      - parameter method: Type of HTTP Method
      - parameter rateLimited: Whether or not the HTTP request needs to be rate limited
      - parameter reason: Optional for when user wants to specify audit-log reason
-    */
+     */
     func requestWithBodyAsData(
-      _ endpoint: Endpoint,
-      params: [String: Any]? = nil,
-      body: Data? = nil,
-      file: String? = nil,
-      authorization: Bool = true,
-      rateLimited: Bool = true,
-      reason: String? = nil,
-      then completion: @escaping (Any?, RequestError?) -> Void
-    ) {
-      let sema = DispatchSemaphore(value: 0)
+        _ endpoint: Endpoint,
+        params: [String: Any]? = nil,
+        body: Data? = nil,
+        file: String? = nil,
+        authorization: Bool = true,
+        rateLimited: Bool = true,
+        reason: String? = nil
+    ) async throws -> Any? {
+        let endpointInfo = endpoint.httpInfo
 
-      let endpointInfo = endpoint.httpInfo
+        var route = self.getRoute(for: endpointInfo.url)
 
-      var route = self.getRoute(for: endpointInfo.url)
-
-      if route.hasSuffix("/messages/:id") && endpointInfo.method == .delete {
-        route += ".delete"
-      }
-
-      var urlString = "https://discord.com/api/v9\(endpointInfo.url)"
-
-      if let params = params {
-        urlString += "?"
-        urlString += params.map({ key, value in "\(key)=\(value)" }
-        ).joined(separator: "&")
-      }
-
-      guard let url = URL(string: urlString) else {
-        self.error(
-          "[Swiftcord] Used an invalid URL: \"\(urlString)\". Please report this."
-        )
-        return
-      }
-
-      var request = URLRequest(url: url)
-
-      request.httpMethod = endpointInfo.method.rawValue
-
-      if authorization {
-        if self.options.isBot {
-          request.addValue("Bot \(token)", forHTTPHeaderField: "Authorization")
-        } else {
-          request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-        }
-      }
-
-      if let reason = reason {
-        request.addValue(
-          reason.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!,
-          forHTTPHeaderField: "X-Audit-Log-Reason"
-        )
-      }
-
-      request.addValue(
-        "DiscordBot (https://github.com/SketchMaster2001/Swiftcord, 0.9.0)",
-        forHTTPHeaderField: "User-Agent"
-      )
-
-      if let body = body {
-        request.httpBody = body
-
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-      }
-
-      let task = self.session.dataTask(with: request) {
-        [unowned self, unowned sema] data, response, error in
-
-        let response = response as! HTTPURLResponse
-        let headers = response.allHeaderFields
-
-        if error != nil {
-          #if !os(Linux)
-          completion(nil, RequestError(error! as NSError))
-          #else
-          completion(nil, RequestError(error as! NSError))
-          #endif
-          sema.signal()
-          return
+        if route.hasSuffix("/messages/:id") && endpointInfo.method == .delete {
+            route += ".delete"
         }
 
-        if rateLimited {
-          self.handleRateLimitHeaders(
-            headers["x-ratelimit-limit"],
-            headers["x-ratelimit-reset"],
-            (headers["Date"] as! String).httpDate.timeIntervalSince1970,
-            route
-          )
+        var urlString = "https://discord.com/api/v9\(endpointInfo.url)"
+
+        if let params = params {
+            urlString += "?"
+            urlString += params.map({ key, value in "\(key)=\(value)" }
+            ).joined(separator: "&")
         }
 
-        if response.statusCode == 204 {
-          completion(nil, nil)
-          sema.signal()
-          return
-        }
-
-        let returnedData = try? JSONSerialization.jsonObject(
-          with: data!,
-          options: .allowFragments
-        )
-
-        if response.statusCode != 200 && response.statusCode != 201 {
-
-          if response.statusCode == 429 {
-            print(
-              "[Swiftcord] You're being rate limited. (This shouldn't happen, check your system clock)"
+        guard let url = URL(string: urlString) else {
+            self.error(
+                "Used an invalid URL: \"\(urlString)\". Please report this."
             )
+            return nil
+        }
 
-            let retryAfter = Int(headers["retry-after"] as! String)!
-            let global = headers["x-ratelimit-global"] as? Bool
+        var request = URLRequest(url: url)
 
-            guard global == nil else {
-              self.isGloballyLocked = true
+        request.httpMethod = endpointInfo.method.rawValue
+
+        if authorization {
+            if self.options.isBot {
+                request.addValue("Bot \(token)", forHTTPHeaderField: "Authorization")
+            } else {
+                request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+            }
+        }
+
+        if let reason = reason {
+            request.addValue(
+                reason.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!,
+                forHTTPHeaderField: "X-Audit-Log-Reason"
+            )
+        }
+
+        request.addValue(
+            "DiscordBot (https://github.com/SketchMaster2001/Swiftcord, 1.0.0)",
+            forHTTPHeaderField: "User-Agent"
+        )
+
+        if let body = body {
+            request.httpBody = body
+
+            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
+        }
+
+        var returnData: Any?
+        returnData = try await withCheckedThrowingContinuation({ continuation in
+            self.baseRequestWithData(
+                request,
+                endpoint,
+                route: route,
+                params: params,
+                body: body,
+                file: file,
+                authorization: authorization,
+                rateLimited: rateLimited,
+                reason: reason
+            ) { data, err in
+                if let err = err {
+                    continuation.resume(throwing: err)
+                } else {
+                    continuation.resume(returning: data)
+                }
+            }
+        })
+
+        return returnData
+    }
+
+    private func baseRequest
+    (
+        _ request: URLRequest,
+        _ endpoint: Endpoint,
+        route: String,
+        params: [String: Any]? = nil,
+        body: [String: Any]? = nil,
+        file: String? = nil,
+        authorization: Bool = true,
+        rateLimited: Bool = true,
+        reason: String? = nil,
+        completion: @escaping (Any?, ResponseError?) -> Void
+    ) {
+        let sema = DispatchSemaphore(value: 0)
+
+        let task = self.session.dataTask(with: request) {
+          [unowned self, unowned sema] data, response, error in
+
+            if error != nil {
+              #if !os(Linux)
+              completion(nil, ResponseError.nonSuccessfulRequest(RequestError(error! as NSError)))
+              #else
+              completion(nil, ResponseError.nonSuccessfulRequest(RequestError(error as! NSError)))
+              #endif
+              sema.signal()
+              return
+            }
+            
+          let response = response as! HTTPURLResponse
+          let headers = response.allHeaderFields
+
+            if response.statusCode == 401 {
+                self.error("Bot token invalid.")
+                completion(nil, ResponseError.nonSuccessfulRequest(RequestError("Bot Token Invalid.")))
+                return
+            }
+
+          if rateLimited {
+            self.handleRateLimitHeaders(
+              headers["x-ratelimit-limit"],
+              headers["x-ratelimit-reset"],
+              (headers["Date"] as! String).httpDate.timeIntervalSince1970,
+              route
+            )
+          }
+
+          if response.statusCode == 204 {
+            completion(nil, nil)
+            sema.signal()
+            return
+          }
+
+          let returnedData = try? JSONSerialization.jsonObject(
+            with: data!,
+            options: .allowFragments
+          )
+
+          if response.statusCode != 200 && response.statusCode != 201 {
+
+            if response.statusCode == 429 {
+                self.warn("You're being rate limited. (This shouldn't happen, check your system clock)")
+
+              let retryAfter = Int(headers["retry-after"] as! String)!
+              let global = headers["x-ratelimit-global"] as? Bool
+
+              guard global == nil else {
+                self.isGloballyLocked = true
+                self.globalQueue.asyncAfter(
+                  deadline: DispatchTime.now() + .seconds(retryAfter)
+                ) { [unowned self] in
+                  self.globalUnlock()
+                }
+
+                sema.signal()
+                return
+              }
+
               self.globalQueue.asyncAfter(
                 deadline: DispatchTime.now() + .seconds(retryAfter)
               ) { [unowned self] in
-                self.globalUnlock()
+                  Task {
+                      try! await self.request(
+                        endpoint,
+                        body: body,
+                        file: file,
+                        authorization: authorization,
+                        rateLimited: rateLimited
+                      )
+                  }
+              }
+            }
+
+            if response.statusCode >= 500 {
+              self.globalQueue.asyncAfter(
+                deadline: DispatchTime.now() + .seconds(3)
+              ) { [unowned self] in
+                  Task {
+                      try! await self.request(
+                        endpoint,
+                        body: body,
+                        file: file,
+                        authorization: authorization,
+                        rateLimited: rateLimited
+                      )
+                  }
               }
 
               sema.signal()
               return
             }
 
-            self.globalQueue.asyncAfter(
-              deadline: DispatchTime.now() + .seconds(retryAfter)
-            ) { [unowned self] in
-              self.requestWithBodyAsData(
-                endpoint,
-                body: body,
-                file: file,
-                authorization: authorization,
-                rateLimited: rateLimited,
-                then: completion
-              )
-            }
-          }
-
-          if response.statusCode >= 500 {
-            self.globalQueue.asyncAfter(
-              deadline: DispatchTime.now() + .seconds(3)
-            ) { [unowned self] in
-              self.requestWithBodyAsData(
-                endpoint,
-                body: body,
-                file: file,
-                authorization: authorization,
-                rateLimited: rateLimited,
-                then: completion
-              )
-            }
-
+            completion(nil, ResponseError.nonSuccessfulRequest(RequestError(response.statusCode, returnedData!)))
             sema.signal()
             return
           }
 
-          completion(nil, RequestError(response.statusCode, returnedData!))
+          completion(returnedData, nil)
+
           sema.signal()
-          return
         }
 
-        completion(returnedData, nil)
+        let apiCall = { [unowned self] in
+          guard rateLimited, self.rateLimits[route] != nil else {
+            task.resume()
 
-        sema.signal()
-      }
+            sema.wait()
+            return
+          }
 
-      let apiCall = { [unowned self] in
-        guard rateLimited, self.rateLimits[route] != nil else {
-          task.resume()
+          let item = DispatchWorkItem {
+            task.resume()
 
-          sema.wait()
-          return
+            sema.wait()
+          }
+
+          self.rateLimits[route]!.queue(item)
         }
 
-        let item = DispatchWorkItem {
-          task.resume()
-
-          sema.wait()
+        if !self.isGloballyLocked {
+          apiCall()
+        } else {
+          self.globalRequestQueue.append(apiCall)
         }
-
-        self.rateLimits[route]!.queue(item)
-      }
-
-      if !self.isGloballyLocked {
-        apiCall()
-      } else {
-        self.globalRequestQueue.append(apiCall)
-      }
 
     }
 
+    private func baseRequestWithData
+    (
+        _ request: URLRequest,
+        _ endpoint: Endpoint,
+        route: String,
+        params: [String: Any]? = nil,
+        body: Data? = nil,
+        file: String? = nil,
+        authorization: Bool = true,
+        rateLimited: Bool = true,
+        reason: String? = nil,
+        completion: @escaping (Any?, ResponseError?) -> Void
+    ) {
+        let sema = DispatchSemaphore(value: 0)
+
+        let task = self.session.dataTask(with: request) {
+          [unowned self, unowned sema] data, response, error in
+
+          let response = response as! HTTPURLResponse
+          let headers = response.allHeaderFields
+
+          if error != nil {
+            #if !os(Linux)
+            completion(nil, ResponseError.nonSuccessfulRequest(RequestError(error! as NSError)))
+            #else
+            completion(nil, ResponseError.nonSuccessfulRequest(RequestError(error as! NSError)))
+            #endif
+            sema.signal()
+            return
+          }
+
+          if rateLimited {
+            self.handleRateLimitHeaders(
+              headers["x-ratelimit-limit"],
+              headers["x-ratelimit-reset"],
+              (headers["Date"] as! String).httpDate.timeIntervalSince1970,
+              route
+            )
+          }
+
+          if response.statusCode == 204 {
+            completion(nil, nil)
+            sema.signal()
+            return
+          }
+
+          let returnedData = try? JSONSerialization.jsonObject(
+            with: data!,
+            options: .allowFragments
+          )
+
+          if response.statusCode != 200 && response.statusCode != 201 {
+
+            if response.statusCode == 429 {
+                self.warn("You're being rate limited. (This shouldn't happen, check your system clock)")
+
+              let retryAfter = Int(headers["retry-after"] as! String)!
+              let global = headers["x-ratelimit-global"] as? Bool
+
+              guard global == nil else {
+                self.isGloballyLocked = true
+                self.globalQueue.asyncAfter(
+                  deadline: DispatchTime.now() + .seconds(retryAfter)
+                ) { [unowned self] in
+                  self.globalUnlock()
+                }
+
+                sema.signal()
+                return
+              }
+
+              self.globalQueue.asyncAfter(
+                deadline: DispatchTime.now() + .seconds(retryAfter)
+              ) { [unowned self] in
+                  Task {
+                      try! await self.requestWithBodyAsData(
+                        endpoint,
+                        body: body,
+                        file: file,
+                        authorization: authorization,
+                        rateLimited: rateLimited
+                      )
+                  }
+              }
+            }
+
+            if response.statusCode >= 500 {
+              self.globalQueue.asyncAfter(
+                deadline: DispatchTime.now() + .seconds(3)
+              ) { [unowned self] in
+                  Task {
+                      try! await self.requestWithBodyAsData(
+                        endpoint,
+                        body: body,
+                        file: file,
+                        authorization: authorization,
+                        rateLimited: rateLimited
+                      )
+                  }
+              }
+
+              sema.signal()
+              return
+            }
+
+            completion(nil, ResponseError.nonSuccessfulRequest(RequestError(response.statusCode, returnedData!)))
+            sema.signal()
+            return
+          }
+
+          completion(returnedData, nil)
+
+          sema.signal()
+        }
+
+        let apiCall = { [unowned self] in
+          guard rateLimited, self.rateLimits[route] != nil else {
+            task.resume()
+
+            sema.wait()
+            return
+          }
+
+          let item = DispatchWorkItem {
+            task.resume()
+
+            sema.wait()
+          }
+
+          self.rateLimits[route]!.queue(item)
+        }
+
+        if !self.isGloballyLocked {
+          apiCall()
+        } else {
+          self.globalRequestQueue.append(apiCall)
+        }
+
+    }
 }
