@@ -11,25 +11,27 @@ import Dispatch
 
 /// <3
 extension Gateway {
-  func heartbeat(at interval: Int) {
-    guard self.isConnected else {
-      return
+    func heartbeat(at interval: Int) {
+        guard self.isConnected else {
+            return
+        }
+
+        guard self.acksMissed < 3 else {
+            Task {
+                print("[Swiftcord] Did not receive ACK from server, reconnecting...")
+                await self.reconnect()
+            }
+            return
+        }
+
+        self.acksMissed += 1
+
+        self.send(self.heartbeatPayload.encode(), presence: false)
+
+        self.heartbeatQueue.asyncAfter(
+            deadline: .now() + .milliseconds(interval)
+        ) { [unowned self] in
+            self.heartbeat(at: interval)
+        }
     }
-    
-    guard self.acksMissed < 3 else {
-      print("[Swiftcord] Did not receive ACK from server, reconnecting...")
-      self.reconnect()
-      return
-    }
-    
-    self.acksMissed += 1
-    
-    self.send(self.heartbeatPayload.encode(), presence: false)
-    
-    self.heartbeatQueue.asyncAfter(
-      deadline: .now() + .milliseconds(interval)
-    ) { [unowned self] in
-      self.heartbeat(at: interval)
-    }
-  }
 }
