@@ -9,44 +9,44 @@
 /// GroupDM Type
 public struct GroupDM: TextChannel {
 
-  // MARK: Properties
+    // MARK: Properties
 
-  /// Parent class
-  public internal(set) weak var swiftcord: Swiftcord?
+    /// Parent class
+    public internal(set) weak var swiftcord: Swiftcord?
 
-  /// ID of DM
-  public let id: Snowflake
+    /// ID of DM
+    public let id: Snowflake
 
-  /// The recipient of this DM
-  public internal(set) var recipients = [User]()
+    /// The recipient of this DM
+    public internal(set) var recipients = [User]()
 
-  /// The last message's ID
-  public let lastMessageId: Snowflake?
+    /// The last message's ID
+    public let lastMessageId: Snowflake?
 
-  /// Indicates what kind of channel this is
-  public let type = ChannelType.groupDM
+    /// Indicates what kind of channel this is
+    public let type = ChannelType.groupDM
 
-  // MARK: Initializer
+    // MARK: Initializer
 
-  /**
-   Creates a GroupChannel struct
+    /**
+     Creates a GroupChannel struct
 
-   - parameter swiftcord: Parent class
-   - parameter json: JSON representable as a dictionary
-   */
-  init(_ swiftcord: Swiftcord, _ json: [String: Any]) {
-    self.swiftcord = swiftcord
+     - parameter swiftcord: Parent class
+     - parameter json: JSON representable as a dictionary
+     */
+    init(_ swiftcord: Swiftcord, _ json: [String: Any]) {
+        self.swiftcord = swiftcord
 
-    self.id = Snowflake(json["id"])!
+        self.id = Snowflake(json["id"])!
 
-    let recipients = json["recipients"] as! [[String: Any]]
-    for recipient in recipients {
-      self.recipients.append(User(swiftcord, recipient))
+        let recipients = json["recipients"] as! [[String: Any]]
+        for recipient in recipients {
+            self.recipients.append(User(swiftcord, recipient))
+        }
+
+        self.lastMessageId = Snowflake(json["last_message_id"])
+
+        swiftcord.groups[self.id] = self
     }
-
-    self.lastMessageId = Snowflake(json["last_message_id"])
-
-    swiftcord.groups[self.id] = self
-  }
 
 }
