@@ -17,11 +17,16 @@ struct InteractionBody<T: Component>: Encodable {
     let embeds: [EmbedBuilder]?
     let flags: Int?
     let components: [ActionRow<T>]?
-    
-    init(content: String? = nil, flags: Int? = nil, embeds: [EmbedBuilder]? = nil, components: [ActionRow<T>]? = nil) {
+
+    public let customId: String?
+    public let title: String?
+
+    init(content: String? = nil, flags: Int? = nil, embeds: [EmbedBuilder]? = nil, components: [ActionRow<T>]? = nil, modal: Modal? = nil) {
         self.content = content
         self.flags = flags
         self.components = components
         self.embeds = embeds
+        self.customId = modal?.customId
+        self.title = modal?.title
     }
 }
