@@ -8,67 +8,92 @@
 import Foundation
 
 public class SlashCommandEvent: InteractionEvent {
-    
+
     public var channelId: Snowflake
-    
-    private let data: [String : Any]
-    
+
+    private let data: [String: Any]
+
     public let interactionId: Snowflake
-    
+
     public let swiftcord: Swiftcord
-    
+
     public let token: String
-    
+
     /// Guild object for this channel
-    public var guild: Guild {
-      return self.swiftcord.getGuild(for: channelId)!
+    public var guild: Guild? {
+        return self.swiftcord.getGuild(for: channelId)
     }
-        
+
     public let name: String
-    
+
     public var member: Member?
-    
+
     public let user: User
-    
+
     public var options = [SlashCommandEventOptions]()
-    
+
     public var ephemeral: Int
-    
+
     public var isDefered: Bool
-    
-    init(_ swiftcord: Swiftcord, data: [String : Any]) {
+
+    init(_ swiftcord: Swiftcord, data: [String: Any]) {
         // Store the data for later
         self.data = data
-        
-        let options = data["data"] as! [String:Any]
-        
+
+        let options = data["data"] as! [String: Any]
+
         self.name = options["name"] as! String
-        
-        for option in options["options"] as! [Any] {
-            self.options.append(SlashCommandEventOptions(data: option as! [String:Any]))
+
+        if let eventOptions =  options["options"] as? [Any] {
+            for option in eventOptions {
+                self.options.append(SlashCommandEventOptions(data: option as! [String: Any]))
+            }
         }
-        
+
         self.channelId = Snowflake(data["channel_id"])!
-        
-        var userJson = data["member"] as! [String:Any]
-        userJson = userJson["user"] as! [String:Any]
+
+        var userJson = data["member"] as! [String: Any]
+        userJson = userJson["user"] as! [String: Any]
         self.user = User(swiftcord, userJson)
-        
+
         self.swiftcord = swiftcord
         self.token = data["token"] as! String
-                
+
         self.interactionId = Snowflake(data["id"] as! String)!
-        
+
         self.ephemeral = 0
         self.isDefered = false
-                
-        self.member = Member(swiftcord, guild, data["member"] as! [String:Any])
+
+        self.member = nil
+        if let guild = guild {
+            self.member = Member(swiftcord, guild, data["member"] as! [String: Any])
+        }
     }
     
+    public func getOptionAsAttachment(optionName: String) -> Attachment? {
+        let options = self.data["data"] as! [String: Any]
+        var attachment: Attachment?
+        
+        if let optionData = options["resolved"] as! [String: Any]? {
+            for option in self.options {
+                if option.name == optionName {
+                    if option.type == .attatchment {
+                        let id = (options["options"] as! [Any])[0] as! [String:Any]
+                        let data = (optionData["attachments"] as! [String:Any])[id["value"] as! String] as! [String:Any]
+                        attachment = Attachment(data)
+                        break
+                    }
+                }
+            }
+        }
+        
+        return attachment
+    }
+
     /// Returns the option data as a `Bool`.
     public func getOptionAsBool(optionName: String) -> Bool? {
-        var bool: Bool? = nil
-        
+        var bool: Bool?
+
         for option in self.options {
             if option.name == optionName {
                 if option.type == .bool {
@@ -80,23 +105,23 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return bool
     }
-    
+
     /// Returns the data from the given option in the Channel protocol. The developer must cast this value to the absolute type.
     public func getOptionAsChannel(optionName: String) -> Channel? {
-        let options = self.data["data"] as! [String:Any]
-        
+        let options = self.data["data"] as! [String: Any]
+
         // Options can be nil, as such we return an optional value
-        var channel: Channel? = nil
-        
-        if let optionData = options["resolved"] as! [String:Any]? {
+        var channel: Channel?
+
+        if let optionData = options["resolved"] as! [String: Any]? {
             for option in self.options {
                 if option.name == optionName {
                     if option.type == .channel {
-                        var channelDict = optionData["channels"] as! [String:Any]
-                        channelDict = channelDict[option.value] as! [String:Any]
+                        var channelDict = optionData["channels"] as! [String: Any]
+                        channelDict = channelDict[option.value] as! [String: Any]
                         channel = GuildText(self.swiftcord, channelDict)
                     } else {
                         self.swiftcord.warn("The option \(optionName) is not a channel!")
@@ -105,14 +130,14 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return channel
     }
-    
+
     /// Returns the option data as a `Double`. Discord considers this type as a `Number`
     public func getOptionAsDouble(optionName: String) -> Double? {
-        var double: Double? = nil
-        
+        var double: Double?
+
         for option in self.options {
             if option.name == optionName {
                 if option.type == .number {
@@ -120,14 +145,14 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return double
     }
-    
+
     /// Returns the option data as an `Int`.
     public func getOptionAsInt(optionName: String) -> Int? {
-        var int: Int? = nil
-        
+        var int: Int?
+
         for option in self.options {
             if option.name == optionName {
                 if option.type == .int {
@@ -135,45 +160,45 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return int
     }
-    
+
     /// Returns the option data as `Member` object.
     public func getOptionAsMember(optionName: String) -> Member? {
-        let options = self.data["data"] as! [String:Any]
-        
-        var member: Member? = nil
-        
-        if let optionData = options["resolved"] as! [String:Any]? {
+        let options = self.data["data"] as! [String: Any]
+
+        var member: Member?
+
+        if let optionData = options["resolved"] as! [String: Any]? {
             for option in self.options {
                 if option.name == optionName {
                     if option.type == .user {
-                        var memberDict = optionData["members"] as! [String:Any]
-                        memberDict = memberDict[option.value] as! [String:Any]
+                        var memberDict = optionData["members"] as! [String: Any]
+                        memberDict = memberDict[option.value] as! [String: Any]
                         member = Member(self.swiftcord, self.swiftcord.getGuild(for: Snowflake(data["channel_id"] as! String)!)!, memberDict)
                         break
                     }
                 }
             }
         }
-        
+
         return member
     }
-    
+
     /// Returns the option data as a `Role` object.
     public func getOptionAsRole(optionName: String) -> Role? {
-        let options = self.data["data"] as! [String:Any]
-        
+        let options = self.data["data"] as! [String: Any]
+
         // Options can be nil, as such we return an optional value
-        var role: Role? = nil
-        
-        if let optionData = options["resolved"] as! [String:Any]? {
+        var role: Role?
+
+        if let optionData = options["resolved"] as! [String: Any]? {
             for option in self.options {
                 if option.name == optionName {
                     if option.type == .role {
-                        var roleDict = optionData["roles"] as! [String:Any]
-                        roleDict = roleDict[option.value] as! [String:Any]
+                        var roleDict = optionData["roles"] as! [String: Any]
+                        roleDict = roleDict[option.value] as! [String: Any]
                         role = Role(roleDict)
                     } else {
                         self.swiftcord.warn("The option \(optionName) is not a role!")
@@ -182,14 +207,14 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return role
     }
-    
+
     /// Returns the option data as a `String`.
-    public func getOptionAsString(optionName: String) -> String? {        
-        var string: String? = nil
-        
+    public func getOptionAsString(optionName: String) -> String? {
+        var string: String?
+
         for option in self.options {
             if option.name == optionName {
                 if option.type == .string {
@@ -197,28 +222,28 @@ public class SlashCommandEvent: InteractionEvent {
                 }
             }
         }
-        
+
         return string
     }
-    
+
     public func getOptionAsUser(optionName: String) -> User? {
-        let options = self.data["data"] as! [String:Any]
-        
-        var user: User? = nil
-        
-        if let optionData = options["resolved"] as! [String:Any]? {
+        let options = self.data["data"] as! [String: Any]
+
+        var user: User?
+
+        if let optionData = options["resolved"] as! [String: Any]? {
             for option in self.options {
                 if option.name == optionName {
                     if option.type == .user {
-                        var userDict = optionData["users"] as! [String:Any]
-                        userDict = userDict[option.value] as! [String:Any]
+                        var userDict = optionData["users"] as! [String: Any]
+                        userDict = userDict[option.value] as! [String: Any]
                         user = User(self.swiftcord, userDict)
                         break
                     }
                 }
             }
         }
-        
+
         return user
     }
 }
@@ -227,18 +252,16 @@ public struct SlashCommandEventOptions {
     public let name: String
     public let type: ApplicationCommandType
     public let value: String
-    
-    init(data: [String:Any]) {
+
+    init(data: [String: Any]) {
         self.name = data["name"] as! String
         self.type = ApplicationCommandType(rawValue: data["type"] as! Int)!
-        
+
         if self.type == .int {
             self.value = String(data["value"] as! Int)
-        }
-        else if self.type == .bool {
+        } else if self.type == .bool {
             self.value = String(data["value"] as! Bool)
-        }
-        else {
+        } else {
             self.value = data["value"] as! String
         }
     }
