@@ -22,13 +22,12 @@ public class SelectMenuBuilder: Encodable {
     }
 }
 
-
 public struct SelectMenu: Component {
     public let type: ComponentTypes
     public let customId: String
     public let options: [SelectMenuOptions]
     public let placeholder: String?
-    
+
     public init(customId: String, placeholder: String? = nil, options: SelectMenuOptions...) {
         self.type = .selectMenu
         self.customId = customId
@@ -42,7 +41,7 @@ public struct SelectMenuOptions: Encodable {
     public let value: String
     public let description: String?
     public let emoji: Emoji?
-    
+
     public init(label: String, value: String, description: String? = nil, emoji: Emoji? = nil) {
         self.label = label
         self.value = value
