@@ -11,33 +11,67 @@ import Foundation
 public struct Sticker: Codable {
     /// Description of the sticker
     public let description: String?
-    
+
     /// ID of the sticker
     public let id: Snowflake?
-    
+
     /// If the sticker is available to use. May be false if the guild has lost boosts
     public let isAvailable: Bool?
-    
+
     /// Name of the sticker
     public let name: String
-    
+
     /// For non-guild stickers, the pack the sticker is from
     public let packId: Snowflake?
-    
+
     /// File format of the sticker
-    public let format: StickerFormat
-    
+    public let format: StickerFormat?
+
+    /// Autocompletion tags for the sticker
+    public let tags: String?
+
     /// Type of sticker
-    public let type: StickerTypes
-    
-    init(_ json: [String:Any]) {
+    public let type: StickerTypes?
+
+    // MARK: Initializer
+
+    /**
+     Creates a Sticker structure from the gateway response
+
+     - parameter json: JSON representable as a dictionary
+     */
+    init(_ json: [String: Any]) {
         self.description = json["description"] as? String
         self.id = Snowflake(json["id"])
         self.isAvailable = json["available"] as? Bool
         self.name = json["name"] as! String
         self.packId = Snowflake(json["pack_id"])
-        self.format = StickerFormat(rawValue: json["format_type"] as! Int)!
-        self.type = StickerTypes(rawValue: json["type"] as! Int)!
+        self.format = StickerFormat(rawValue: json["format_type"] as! Int)
+        self.tags = json["tags"] as? String
+        self.type = StickerTypes(rawValue: json["type"] as! Int)
+    }
+
+    /**
+     Creates a Sticker structure for uploading or editing
+
+     - parameter name: Name of the sticker
+     - parameter description: Description of the sticker
+     - parameter tags: A string formatted like comma-seperated values for autocompletion
+     */
+    public init(
+        name: String,
+        description: String,
+        tags: String
+    ) {
+        self.name = name
+        self.description = description
+        self.tags = tags
+
+        self.id = nil
+        self.isAvailable = nil
+        self.packId = nil
+        self.format = nil
+        self.type = nil
     }
 }
 
@@ -45,7 +79,7 @@ public struct Sticker: Codable {
 public enum StickerTypes: Int, Codable {
     /// An official sticker made by Discord
     case standard = 1
-    
+
     /// Stickers found in a guild
     case guild
 }
