@@ -2,10 +2,37 @@
 import XCTest
 
 final class SwiftCordTests: XCTestCase {
+    class TestMessageSend: ListenerAdapter {
+        override func onMessageCreate(event: Message) async {
+            if event.content == "+test" {
+                let msg = try! await event.reply(with: "Testing Swiftcord!")
+                // Do whatever you want with the message object
+                sleep(2)
+                try! await msg!.delete()
+            }
+        }
+    }
+    
+    class TestEmbedSend: ListenerAdapter {
+        override func onMessageCreate(event: Message) async {
+            if event.content == "+embed" {
+                let embed = EmbedBuilder()
+                    .setTitle(title: "Swiftcord Embed")
+                    .setDescription(description: "This embed shows off the embed")
+                    .addField("Field", value: "A field")
+                    .setFooter(text: "Created in Swiftcord")
+                    .setTimestamp()
+                
+                let _ = try! await event.reply(with: embed)
+                // Again, do whatever you want with the message object
+            }
+        }
+    }
+    
     /// Sets up the `Swiftcord` object needed across functions
     func setUpBot() -> Swiftcord {
         let bot = Swiftcord(token: "")
-        bot.setIntents(intents: .guilds, .guildMessages)
+        bot.setIntents(intents: .guildMessages)
         
         let activity = Activities(name: "WiiLink Championships", type: .competing)
 
@@ -21,36 +48,15 @@ final class SwiftCordTests: XCTestCase {
     
     func testMessageCommand() {
         let bot = self.setUpBot()
-        
-        bot.on(.messageCreate) { data in
-            let msg = data as! Message
-            
-            if msg.content == "+test" {
-                msg.reply(with: "Testing Swiftcord!")
-            }
-        }
+        bot.addListeners(TestMessageSend())
         
         bot.connect()
     }
     
     func testSendEmbed() {
         let bot = self.setUpBot()
-        
-        bot.on(.messageCreate) { data in
-            let msg = data as! Message
-            
-            if msg.content == "+test" {
-                let embed = EmbedBuilder()
-                    .setTitle(title: "Swiftcord Embed")
-                    .setDescription(description: "This embed shows off the embed")
-                    .addField("Field", value: "A field")
-                    .setFooter(text: "Created in Swiftcord")
-                    .setTimestamp()
+        bot.addListeners(TestEmbedSend())
                 
-                msg.reply(with: embed)
-            }
-        }
-        
         bot.connect()
     }
 }
